Feat: improving albedo ranges in validator (#978)
* add ge=0 to roof_albedo_dir_mult_fact

In suews_type_spartacus.f95, commented as: Ratio of direct/diffuse roof albedo [line 37]. Should be non-negative.

* ranges for alb_id (vegetated surface initial albedo)

* preciplimitalb must be non-negative

Despite not being directly an albedo, it is still related to snow radiation calculations and needs to be non-negative (as defined in Fortran)

* style: auto-format code with ruff and fprettify

Co-authored-by: dayantur <71443948+dayantur@users.noreply.github.com>

* albedo ranges for air and veg single-scattering longwave and shortwave radiation in SPARTACUS

* ground_albedo_dir_mult_fact constrained to be non-negative

Ratio / multiplier between direct and diffuse components of ground albedo. Must be non negative.

* Update CHANGELOG.md

* add PR number to changelog

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>